+++ kinto/core/listeners/__init__.py
@@ -1,4 +1,3 @@
-
 class ListenerBase(object):
     def __init__(self, *args, **kwargs):
         pass